refactor: LikeToggle로 네이밍 변경
