+++ src/components/molecule/LikeToggle/LikeToggle.stories.jsx
@@ -1,12 +1,12 @@
-import { LikeBox } from '@components';
+import { LikeToggle } from '@components';
 
 export default {
-  title: 'Component/Domain/LikeBox',
+  title: 'Component/Domain/LikeToggle',
 };
 
 export const Default = () => {
   const handleClick = (toggle, count) => {
     console.log('좋아요 클릭', toggle, count);
   };
-  return <LikeBox interactive onClick={handleClick} />;
+  return <LikeToggle interactive onClick={handleClick} />;
 };
